fix(mcp,cli,plans): address remaining coderabbitai comments and complete followups
Security test refactoring (coderabbitai #1):
- Split 827-LOC memory-cli/tests/security_tests.rs into 6 focused submodules
- bounds_tests, config_tests, injection_tests, path_traversal_tests, validation_tests
- Each file well under 500 LOC limit

CURRENT.md fix (coderabbitai on PR #549):
- Add blank lines between headings and tables (MD022/MD058)
- Move PRs #547/#546 to Recently Merged (both merged 2026-05-16)
- List PR #549 as only open PR
- Update test count: 2902 -> 3288

Quality gate fix:
- Trim tag/core.rs from 503 to 500 LOC (remove blank lines + verbose comment)

Dead import cleanup (code review):
- Remove unused imports in injection_tests.rs and mod.rs
