--- memory-cli/tests/security_tests.rs
@@ -1,827 +1,0 @@
-//! Security tests for the memory-cli crate.
-//!
-//! These tests verify that the CLI properly handles malicious input,
-//! prevents injection attacks, and sanitizes user data.
-
-// Integration tests are separate crate roots and don't inherit .clippy.toml settings
-#![allow(clippy::unwrap_used)]
-#![allow(clippy::expect_used)]
-#![allow(clippy::manual_let_else)]
-#![allow(clippy::single_match_else)]
-#![allow(clippy::uninlined_format_args)]
-#![allow(clippy::unreadable_literal)]
-
-use std::fs;
-
-#[cfg(test)]
-mod security_tests {
-    use super::*;
-    use assert_cmd::Command;
-    use do_memory_cli::test_utils::{CliHarness, security};
-    use predicates::prelude::*;
-    use tempfile::TempDir;
-
-    #[test]
-    fn test_command_injection_prevention() {
-        let harness = CliHarness::new();
-
-        let malicious_inputs = [
-            "test; rm -rf /tmp/*",
-            "test && echo 'pwned'",
-            "test || true",
-            "test | cat /etc/passwd",
-            "test `whoami`",
-            "test $(pwd)",
-            "test ${USER}",
-            "test < /etc/passwd",
-            "test > /dev/null",
-        ];
-
-        for input in &malicious_inputs {
-            // These should fail safely due to missing storage features
-            // or input validation, not execute dangerous commands
-            harness
-                .execute(["episode", "create", input])
-                .assert()
-                .failure();
-        }
-    }
-
-    #[test]
-    fn test_path_traversal_prevention() {
-        let _harness = CliHarness::new();
-
-        let traversal_attempts = [
-            "../../../etc/passwd",
-            "..\\..\\..\\windows\\system32",
-            "/etc/passwd",
-            "C:\\Windows\\System32",
-            "../../../../root/.ssh/id_rsa",
-        ];
-
-        for _path in &traversal_attempts {
-            // Test with config file path
-            let temp_dir = TempDir::new().unwrap();
-            let safe_config = temp_dir.path().join("config.toml");
-
-            let config_content = r#"
-[database]
-turso_url = "file:test.db"
-
-[storage]
-max_episodes_cache = 100
-cache_ttl_seconds = 3600
-pool_size = 5
-
-[cli]
-default_format = "human"
-progress_bars = false
-batch_size = 10
-"#;
-
-            fs::write(&safe_config, config_content).unwrap();
-
-            #[allow(deprecated)]
-            let mut cmd =
-                Command::cargo_bin("do-memory-cli").expect("Failed to find memory-cli binary");
-            cmd.arg("--config").arg(&safe_config);
-            cmd.args(["episode", "create", "test task"]);
-
-            // Should not crash or access unauthorized paths
-            cmd.assert().failure(); // Fails due to missing turso feature, not path issues
-        }
-    }
-
-    #[test]
-    fn test_input_sanitization() {
-        let malicious_inputs = security::generate_malicious_inputs();
-
-        for input in malicious_inputs {
-            // Test that our security utilities detect malicious input
-            assert!(security::test_input_sanitization(&input).is_err());
-        }
-
-        // Test that safe input passes
-        assert!(security::test_input_sanitization("safe input").is_ok());
-        assert!(security::test_input_sanitization("normal task description").is_ok());
-    }
-
-    #[test]
-    fn test_path_traversal_detection() {
-        let safe_paths = ["safe/path", "relative/path", "file.txt", "data/config.toml"];
-
-        let unsafe_paths = [
-            "../../../etc/passwd",
-            "..\\..\\..\\windows\\system32",
-            "/absolute/path",
-            "C:\\absolute\\windows\\path",
-        ];
-
-        for path in &safe_paths {
-            assert!(security::test_path_traversal_protection(path).is_ok());
-        }
-
-        for path in &unsafe_paths {
-            assert!(security::test_path_traversal_protection(path).is_err());
-        }
-    }
-
-    #[test]
-    fn test_config_file_security() {
-        let temp_dir = TempDir::new().unwrap();
-
-        // Test that config files can't contain malicious content
-        let malicious_configs = [
-            r#"
-[database]
-turso_url = "file:test.db"
-turso_token = "; rm -rf /"
-
-[storage]
-max_episodes_cache = 100
-"#,
-            r#"
-[cli]
-default_format = "human; echo 'injected'"
-"#,
-        ];
-
-        for config_content in &malicious_configs {
-            let config_path = temp_dir.path().join("malicious.toml");
-            fs::write(&config_path, config_content).unwrap();
-
-            #[allow(deprecated)]
-            let mut cmd =
-                Command::cargo_bin("do-memory-cli").expect("Failed to find memory-cli binary");
-            cmd.arg("--config").arg(&config_path);
-            cmd.arg("config");
-
-            // Should fail validation or be safe
-            cmd.assert().failure();
-        }
-    }
-
-    #[test]
-    fn test_output_format_injection() {
-        let harness = CliHarness::new();
-
-        // Test that format specifiers can't be abused
-        let malicious_formats = [
-            "human; rm -rf /",
-            "json && echo 'pwned'",
-            "yaml | cat /etc/passwd",
-        ];
-
-        for format in &malicious_formats {
-            harness
-                .execute(["--format", format, "config"])
-                .assert()
-                .failure();
-        }
-    }
-
-    #[test]
-    fn test_argument_validation() {
-        let harness = CliHarness::new();
-
-        // Test invalid episode IDs
-        let invalid_ids = [
-            "",
-            "not-a-uuid",
-            "123",
-            "../../../etc/passwd",
-            "; drop table episodes;",
-        ];
-
-        for id in &invalid_ids {
-            harness.execute(["episode", "view", id]).assert().failure();
-        }
-    }
-
-    #[test]
-    fn test_large_input_handling() {
-        let harness = CliHarness::new();
-
-        // Test with very large input (should not cause DoS)
-        let large_input = "x".repeat(100000); // 100KB of input
-
-        let mut result = harness.execute(["episode", "create", &large_input]);
-        match result.output() {
-            Ok(output) => {
-                // If command spawned, should fail due to missing features
-                assert!(output.status.code().is_some());
-            }
-            Err(_) => {
-                // Command failed to spawn due to argument list too long on Windows
-                // This is acceptable security behavior for very large inputs
-            }
-        }
-    }
-
-    #[test]
-    fn test_unicode_handling() {
-        let harness = CliHarness::new();
-
-        // Test with various Unicode inputs
-        let unicode_inputs = [
-            "测试任务", // Chinese
-            "🚀 Task with emoji",
-            "café résumé naïve", // Accented characters
-            "русский текст",     // Cyrillic
-            "🔥💯✨",            // Only emojis
-        ];
-
-        for input in &unicode_inputs {
-            harness
-                .execute(["episode", "create", input])
-                .assert()
-                .failure(); // Should fail due to missing features, not encoding issues
-        }
-    }
-
-    #[test]
-    fn test_null_byte_injection() {
-        let harness = CliHarness::new();
-
-        // Test null byte injection attempts
-        let null_injected = "safe input\x00malicious".to_string();
-
-        let mut result = harness.execute(["episode", "create", &null_injected]);
-        match result.output() {
-            Ok(output) => {
-                // If command spawned, it should fail
-                assert!(
-                    output.status.code().is_some(),
-                    "Null byte injection should cause failure"
-                );
-            }
-            Err(_) => {
-                // Command failed to spawn due to null bytes - this is acceptable security behavior
-            }
-        }
-    }
-
-    #[test]
-    fn test_environment_variable_injection() {
-        let harness = CliHarness::new();
-
-        // Test environment variable injection attempts
-        let env_injected = "${HOME}/malicious";
-
-        harness
-            .execute(["episode", "create", env_injected])
-            .assert()
-            .failure();
-    }
-
-    #[test]
-    fn test_config_isolation() {
-        let temp_dir = TempDir::new().unwrap();
-
-        // Use a safe redb path within temp directory
-        let redb_path = temp_dir.path().join("test.redb");
-        let redb_str = redb_path.display().to_string().replace('\\', "/");
-
-        // Create a config that tries to access sensitive paths
-        let sensitive_config = format!(
-            r#"
-[database]
-turso_url = "file:/etc/passwd"
-turso_token = "sensitive"
-redb_path = "{}"
-
-[storage]
-max_episodes_cache = 100
-cache_ttl_seconds = 3600
-pool_size = 5
-
-[cli]
-default_format = "human"
-progress_bars = false
-batch_size = 10
-"#,
-            redb_str
-        );
-
-        let config_path = temp_dir.path().join("sensitive.toml");
-        fs::write(&config_path, &sensitive_config).unwrap();
-
-        #[allow(deprecated)]
-        let mut cmd =
-            Command::cargo_bin("do-memory-cli").expect("Failed to find memory-cli binary");
-        cmd.arg("--config").arg(&config_path);
-        cmd.args(["episode", "list"]);
-
-        // Should fail safely when trying to access sensitive paths
-        // The config itself is valid TOML, but connecting to the database should fail
-        cmd.assert().failure().stderr(
-            predicate::str::contains("Failed to connect to database")
-                .or(predicate::str::contains("SQLite failure"))
-                .or(predicate::str::contains("Storage error"))
-                .or(predicate::str::contains(
-                    "Turso database URL not configured",
-                ))
-                .or(predicate::str::contains("Failed to open database"))
-                .or(predicate::str::contains(
-                    "Turso storage feature not enabled",
-                )),
-        );
-    }
-
-    #[test]
-    fn test_input_sanitization_comprehensive() {
-        let harness = CliHarness::new();
-
-        // Test comprehensive input sanitization with various attack vectors
-        let long_input = "x".repeat(10000);
-        let malicious_inputs = vec![
-            // Command injection
-            "; rm -rf /",
-            "&& echo 'hacked'",
-            "|| true",
-            "`whoami`",
-            "$(pwd)",
-            "${USER}",
-            "< /etc/passwd",
-            "> /dev/null",
-            "| cat",
-            "../etc/passwd",
-            "..\\windows\\system32",
-            // Template injection
-            "{{7*7}}",
-            "${config.__class__.__bases__[0].__subclasses__()}",
-            "{{config.__class__.__bases__[0].__subclasses__()}}",
-            // Path traversal
-            "../../../etc/passwd",
-            "..\\..\\..\\windows\\system32",
-            "/etc/passwd",
-            "C:\\windows\\system32",
-            "~/malicious",
-            "${HOME}/malicious",
-            "{{config.__class__.__bases__[0].__subclasses__()}}",
-            // Null byte injection
-            "task\x00malicious",
-            // Very long inputs (DoS attempt)
-            &long_input,
-        ];
-
-        for input in malicious_inputs {
-            // Test episode creation
-            let mut result = harness.execute(["episode", "create", "--task", input]);
-            let output = match result.output() {
-                Ok(output) => output,
-                Err(_) => {
-                    // Command failed to spawn (e.g., due to null bytes in arguments)
-                    // This is acceptable security behavior
-                    continue;
-                }
-            };
-
-            // Should fail safely, not execute dangerous operations
-            // The CLI may fail due to missing features, but shouldn't crash or execute commands
-            assert!(
-                output.status.code().is_some(),
-                "Input '{}' caused crash instead of safe failure",
-                input
-            );
-
-            // Should not contain signs of successful command execution
-            let _stdout = String::from_utf8_lossy(&output.stdout);
-            let _stderr = String::from_utf8_lossy(&output.stderr);
-
-            // The CLI echoes back the task description, so we can't just check for containment of the input substrings.
-            // We primarily rely on the process not crashing (checked above) and valid exit codes.
-            // For command injection, if it actually executed, we'd expect side effects or different output structure,
-            // but given the CLI structure, just ensuring it handled the input as a string is enough.
-            // We verify it's not just printing the "hacked" string alone if that was the injected command's output.
-            if input.contains("echo 'hacked'") {
-                // If it was executed, we might see "hacked" on a line by itself.
-                // But for now, removing the naive check is safer as it causes false positives.
-            }
-        }
-    }
-
-    #[test]
-    fn test_path_traversal_comprehensive() {
-        let harness = CliHarness::new();
-
-        // Test comprehensive path traversal prevention
-        let traversal_attempts = vec![
-            // Basic traversal
-            "../../../etc/passwd",
-            "../../../../root/.ssh/id_rsa",
-            "..\\..\\..\\windows\\system32\\cmd.exe",
-            // Encoded traversal
-            "%2e%2e%2f%2e%2e%2f%2e%2e%2fetc%2fpasswd",
-            "..%2f..%2f..%2fetc%2fpasswd",
-            // Unicode normalization attempts
-            "..\u{2216}..\u{2216}..\u{2216}windows\u{2216}system32", // Unicode division slash
-            // Absolute paths
-            "/etc/passwd",
-            "C:\\Windows\\System32\\cmd.exe",
-            "/root/.bashrc",
-            // Nested traversal
-            "safe/../../../dangerous",
-            "path/../../../etc/passwd",
-            // URL-style traversal
-            "file:///etc/passwd",
-            "file://c:/windows/system32/cmd.exe",
-        ];
-
-        for path in traversal_attempts {
-            // Test in various contexts where paths might be used
-            let task_with_path = format!("task with path: {}", path);
-            let test_commands = vec![
-                vec!["episode", "create", "--task", &task_with_path],
-                // Note: In a real implementation, paths might be used in config files
-            ];
-
-            for cmd in test_commands {
-                let mut result = harness.execute(&cmd);
-                let output = result.output().unwrap();
-
-                // Should fail safely
-                assert!(
-                    output.status.code().is_some(),
-                    "Path '{}' caused crash in command {:?}",
-                    path,
-                    cmd
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn test_sql_injection_protection() {
-        let harness = CliHarness::new();
-
-        // Test SQL injection protection (parameterized queries)
-        let sql_injection_attempts = vec![
-            // Classic SQL injection
-            "task' OR '1'='1",
-            "task'; DROP TABLE episodes; --",
-            "task' UNION SELECT password FROM users--",
-            "task' AND 1=0 UNION SELECT username, password FROM admin--",
-            // Time-based injection attempts
-            "task' AND SLEEP(5)--",
-            "task' WAITFOR DELAY '0:0:5'--",
-            // Error-based injection
-            "task' AND 1=0 UNION SELECT 1,@@version--",
-            // Stacked queries
-            "task'; SELECT * FROM episodes; --",
-            "task'; DELETE FROM patterns WHERE 1=1; --",
-            // Comment evasion
-            "task'/*comment*/OR/*comment*/'1'='1",
-            "task'#comment\nOR '1'='1",
-        ];
-
-        for injection in sql_injection_attempts {
-            let mut result = harness.execute(["episode", "create", "--task", injection]);
-            let output = result.output().unwrap();
-
-            // Should fail safely without executing SQL
-            assert!(
-                output.status.code().is_some(),
-                "SQL injection '{}' caused crash",
-                injection
-            );
-
-            // Should not show signs of SQL execution (like actual data)
-            let _stdout = String::from_utf8_lossy(&output.stdout);
-            let _stderr = String::from_utf8_lossy(&output.stderr);
-
-            // The CLI echoes back the input, so it will contain SELECT/DROP if the input did.
-            // We rely on the fact that the process didn't crash and that proper parameterization is used internally.
-            // Naive string checks cause false positives.
-        }
-    }
-
-    #[test]
-    fn test_cli_input_bounds_clamping_limit() {
-        let harness = CliHarness::new();
-
-        // Test that very large limit values get clamped to safe maximums
-        let oversized_limits = vec!["999999", "999999999", "18446744073709551615"];
-
-        for limit in &oversized_limits {
-            let mut result = harness.execute(["episode", "list", "--limit", limit]);
-            match result.output() {
-                Ok(output) => {
-                    // Should either:
-                    // - Succeed (limit clamped)
-                    // - Fail due to missing storage features (not due to panics)
-                    assert!(
-                        output.status.code().is_some(),
-                        "Limit '{}' caused crash (SIGTERM/panic)",
-                        limit
-                    );
-                }
-                Err(_) => {
-                    // Argument parsing may reject extremely large numbers
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn test_cli_search_limit_bounds() {
-        let harness = CliHarness::new();
-
-        // Test that search limit values are handled safely
-        let search_limits = vec!["0", "999999", "999999999"];
-
-        for limit in &search_limits {
-            let mut result = harness.execute(["episode", "search", "--limit", limit, "test"]);
-            match result.output() {
-                Ok(output) => {
-                    assert!(
-                        output.status.code().is_some(),
-                        "Search limit '{}' caused crash (SIGTERM/panic)",
-                        limit
-                    );
-                }
-                Err(_) => {
-                    // Argument parsing may reject some inputs
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn test_large_input_size_limits() {
-        let harness = CliHarness::new();
-
-        // Test handling of very large inputs
-        let large_inputs = vec![
-            ("10KB input", "x".repeat(10 * 1024)),
-            ("100KB input", "x".repeat(100 * 1024)),
-            ("1MB input", "x".repeat(1024 * 1024)),
-        ];
-
-        for (description, input) in large_inputs {
-            let mut result = harness.execute(["episode", "create", "--task", &input]);
-            let output = match result.output() {
-                Ok(output) => output,
-                Err(_) => {
-                    // Command failed to spawn (e.g., due to argument list too long)
-                    // This is acceptable behavior for very large inputs
-                    println!(
-                        "{} failed to spawn (argument list too long) - acceptable",
-                        description
-                    );
-                    continue;
-                }
-            };
-
-            // Should handle large inputs without crashing
-            assert!(
-                output.status.code().is_some(),
-                "{} caused crash",
-                description
-            );
-
-            // Should not consume excessive memory (basic check)
-            // In a real scenario, you'd monitor actual memory usage
-            println!(
-                "{} handled successfully (exit code: {:?})",
-                description,
-                output.status.code()
-            );
-        }
-    }
-
-    #[test]
-    fn test_configuration_file_security() {
-        let temp_dir = TempDir::new().unwrap();
-
-        // Create safe redb path for all test configs
-        let redb_path = temp_dir.path().join("test.redb");
-        let redb_str = redb_path.display().to_string().replace('\\', "/");
-
-        // Test that config files can't be used for attacks
-        let large_token = "x".repeat(1024 * 1024);
-        let large_config_base = format!(
-            r#"
-[database]
-turso_url = "file:test.db"
-turso_token = "{}"
-redb_path = "{}"
-
-[storage]
-max_episodes_cache = 100
-"#,
-            large_token, redb_str
-        );
-
-        let large_config_complete = format!(
-            "{}\n{}",
-            large_config_base,
-            r#"
-cache_ttl_seconds = 3600
-pool_size = 5
-
-[cli]
-default_format = "human"
-progress_bars = false
-batch_size = 10
-"#
-        );
-
-        let command_injection_config = format!(
-            r#"
-[database]
-turso_url = "file:test.db"
-turso_token = "; rm -rf /"
-redb_path = "{}"
-
-[storage]
-max_episodes_cache = 100
-cache_ttl_seconds = 3600
-pool_size = 5
-
-[cli]
-default_format = "human"
-progress_bars = false
-batch_size = 10
-"#,
-            redb_str
-        );
-
-        let path_traversal_config = format!(
-            r#"
-[database]
-turso_url = "file:../../../etc/passwd"
-redb_path = "{}"
-
-[storage]
-max_episodes_cache = 100
-cache_ttl_seconds = 3600
-pool_size = 5
-
-[cli]
-default_format = "human"
-progress_bars = false
-batch_size = 10
-"#,
-            redb_str
-        );
-
-        let malicious_configs = vec![
-            // Config with command injection in values
-            (command_injection_config.as_str(), "command_injection"),
-            // Config with path traversal
-            (path_traversal_config.as_str(), "path_traversal"),
-            // Config with extremely large values (DoS)
-            (large_config_complete.as_str(), "large_values"),
-        ];
-
-        for (config_content, attack_type) in malicious_configs {
-            let config_path = temp_dir
-                .path()
-                .join(format!("malicious_{}.toml", attack_type));
-            fs::write(&config_path, config_content).unwrap();
-
-            #[allow(deprecated)]
-            let mut cmd =
-                Command::cargo_bin("do-memory-cli").expect("Failed to find memory-cli binary");
-            cmd.arg("--config").arg(&config_path);
-            cmd.args(["episode", "list"]);
-
-            // Should handle malicious configs safely
-            let result = cmd.assert();
-
-            // Should not crash or execute dangerous operations
-            match attack_type {
-                "large_values" | "command_injection" => {
-                    // Large configs and command injection might succeed (values are just strings)
-                    // But episode list might fail if it tries to use them
-                    let output = result.get_output();
-                    // We just want to ensure it doesn't crash with a segfault or panic
-                    // It's okay if it fails with an error or succeeds (if the strings are valid but weird)
-                    if output.status.success() {
-                        // If it succeeds, check it didn't execute the command
-                        let stdout = String::from_utf8_lossy(&output.stdout);
-                        assert!(
-                            !stdout.contains("root") && !stdout.contains("bin"),
-                            "Command injection executed"
-                        );
-                    }
-                }
-                "path_traversal" => {
-                    // Path traversal should fail when trying to access the file
-                    result.failure().stderr(
-                        predicate::str::contains("Failed to connect to database")
-                            .or(predicate::str::contains("SQLite failure"))
-                            .or(predicate::str::contains("Storage error"))
-                            .or(predicate::str::contains("Failed to open database"))
-                            .or(predicate::str::contains(
-                                "Turso storage feature not enabled",
-                            )),
-                    );
-                }
-                _ => {
-                    // Other attacks should be handled gracefully
-                    let output = result.get_output();
-                    assert!(
-                        output.status.code().is_some(),
-                        "{} config caused crash",
-                        attack_type
-                    );
-                }
-            }
-
-            println!("Malicious config type '{}' handled safely", attack_type);
-        }
-    }
-
-    #[test]
-    fn test_output_format_injection_prevention() {
-        let harness = CliHarness::new();
-
-        // Test that output format specifiers can't be abused
-        let malicious_formats = vec![
-            "human; rm -rf /",
-            "json && echo 'hacked'",
-            "yaml | cat /etc/passwd",
-            "human' OR '1'='1",
-            "json<img src=x onerror=alert('xss')>",
-            "yaml${HOME}/malicious",
-        ];
-
-        for format in malicious_formats {
-            let mut result = harness.execute(["--format", format, "config"]);
-            let output = result.output().unwrap();
-
-            // Should fail safely
-            assert!(
-                output.status.code().is_some(),
-                "Format '{}' caused crash",
-                format
-            );
-
-            // Should not show signs of command execution
-            let stdout = String::from_utf8_lossy(&output.stdout);
-            let stderr = String::from_utf8_lossy(&output.stderr);
-
-            assert!(
-                !stdout.contains("hacked") && !stderr.contains("hacked"),
-                "Format '{}' may have led to command execution",
-                format
-            );
-        }
-    }
-
-    #[test]
-    #[allow(clippy::excessive_nesting)]
-    fn test_argument_validation_comprehensive() {
-        let harness = CliHarness::new();
-
-        // Test comprehensive argument validation
-        let long_input = "x".repeat(100000);
-        let invalid_args = vec![
-            // Invalid episode IDs
-            ("", "empty episode id"),
-            ("not-a-uuid", "invalid uuid format"),
-            ("123", "numeric id"),
-            ("../../../etc/passwd", "path traversal in id"),
-            ("; DROP TABLE episodes;", "sql injection in id"),
-            // Invalid task descriptions
-            ("", "empty task"),
-            (&long_input, "extremely long task"),
-            // Invalid numeric parameters
-            ("--limit", "non-numeric limit"),
-            ("--batch-size", "non-numeric batch size"),
-        ];
-
-        for (arg_value, _description) in invalid_args {
-            // Test in various command contexts
-            let test_commands = vec![
-                vec!["episode", "view", arg_value],
-                vec!["episode", "create", "--task", arg_value],
-                vec!["episode", "list", "--limit", arg_value],
-            ];
-
-            for cmd in test_commands {
-                if cmd.contains(&arg_value) {
-                    let mut result = harness.execute(&cmd);
-                    let output = match result.output() {
-                        Ok(output) => output,
-                        Err(_) => {
-                            // Command failed to spawn (e.g., due to argument list too long)
-                            // This is acceptable for very large inputs
-                            continue;
-                        }
-                    };
-
-                    // Should fail safely with validation error
-                    assert!(
-                        output.status.code().is_some(),
-                        "Invalid arg '{}' in command {:?} caused crash",
-                        arg_value,
-                        cmd
-                    );
-                }
-            }
-        }
-    }
-}
