+++ memory-cli/tests/security_tests/bounds_tests.rs
@@ -1,0 +1,122 @@
+//! Input bounds and size limit tests.
+
+use do_memory_cli::test_utils::CliHarness;
+
+#[cfg(test)]
+mod bounds_tests {
+    use super::*;
+
+    #[test]
+    fn test_cli_input_bounds_clamping_limit() {
+        let harness = CliHarness::new();
+
+        // Test that very large limit values get clamped to safe maximums
+        let oversized_limits = vec!["999999", "999999999", "18446744073709551615"];
+
+        for limit in &oversized_limits {
+            let mut result = harness.execute(["episode", "list", "--limit", limit]);
+            match result.output() {
+                Ok(output) => {
+                    // Should either:
+                    // - Succeed (limit clamped)
+                    // - Fail due to missing storage features (not due to panics)
+                    assert!(
+                        output.status.code().is_some(),
+                        "Limit '{}' caused crash (SIGTERM/panic)",
+                        limit
+                    );
+                }
+                Err(_) => {
+                    // Argument parsing may reject extremely large numbers
+                }
+            }
+        }
+    }
+
+    #[test]
+    fn test_cli_search_limit_bounds() {
+        let harness = CliHarness::new();
+
+        // Test that search limit values are handled safely
+        let search_limits = vec!["0", "999999", "999999999"];
+
+        for limit in &search_limits {
+            let mut result = harness.execute(["episode", "search", "--limit", limit, "test"]);
+            match result.output() {
+                Ok(output) => {
+                    assert!(
+                        output.status.code().is_some(),
+                        "Search limit '{}' caused crash (SIGTERM/panic)",
+                        limit
+                    );
+                }
+                Err(_) => {
+                    // Argument parsing may reject some inputs
+                }
+            }
+        }
+    }
+
+    #[test]
+    fn test_large_input_handling() {
+        let harness = CliHarness::new();
+
+        // Test with very large input (should not cause DoS)
+        let large_input = "x".repeat(100000); // 100KB of input
+
+        let mut result = harness.execute(["episode", "create", &large_input]);
+        match result.output() {
+            Ok(output) => {
+                // If command spawned, should fail due to missing features
+                assert!(output.status.code().is_some());
+            }
+            Err(_) => {
+                // Command failed to spawn due to argument list too long on Windows
+                // This is acceptable security behavior for very large inputs
+            }
+        }
+    }
+
+    #[test]
+    fn test_large_input_size_limits() {
+        let harness = CliHarness::new();
+
+        // Test handling of very large inputs
+        let large_inputs = vec![
+            ("10KB input", "x".repeat(10 * 1024)),
+            ("100KB input", "x".repeat(100 * 1024)),
+            ("1MB input", "x".repeat(1024 * 1024)),
+        ];
+
+        for (description, input) in large_inputs {
+            let mut result = harness.execute(["episode", "create", "--task", &input]);
+            let output = match result.output() {
+                Ok(output) => output,
+                Err(_) => {
+                    // Command failed to spawn (e.g., due to argument list too long)
+                    // This is acceptable behavior for very large inputs
+                    println!(
+                        "{} failed to spawn (argument list too long) - acceptable",
+                        description
+                    );
+                    continue;
+                }
+            };
+
+            // Should handle large inputs without crashing
+            assert!(
+                output.status.code().is_some(),
+                "{} caused crash",
+                description
+            );
+
+            // Should not consume excessive memory (basic check)
+            // In a real scenario, you'd monitor actual memory usage
+            println!(
+                "{} handled successfully (exit code: {:?})",
+                description,
+                output.status.code()
+            );
+        }
+    }
+}
