+++ memory-cli/tests/security_tests/injection_tests.rs
@@ -1,0 +1,247 @@
+//! Injection attack prevention tests.
+
+use do_memory_cli::test_utils::{CliHarness, security};
+
+#[cfg(test)]
+mod injection_tests {
+    use super::*;
+
+    #[test]
+    fn test_command_injection_prevention() {
+        let harness = CliHarness::new();
+
+        let malicious_inputs = [
+            "test; rm -rf /tmp/*",
+            "test && echo 'pwned'",
+            "test || true",
+            "test | cat /etc/passwd",
+            "test `whoami`",
+            "test $(pwd)",
+            "test ${USER}",
+            "test < /etc/passwd",
+            "test > /dev/null",
+        ];
+
+        for input in &malicious_inputs {
+            // These should fail safely due to missing storage features
+            // or input validation, not execute dangerous commands
+            harness
+                .execute(["episode", "create", input])
+                .assert()
+                .failure();
+        }
+    }
+
+    #[test]
+    fn test_input_sanitization() {
+        let malicious_inputs = security::generate_malicious_inputs();
+
+        for input in malicious_inputs {
+            // Test that our security utilities detect malicious input
+            assert!(security::test_input_sanitization(&input).is_err());
+        }
+
+        // Test that safe input passes
+        assert!(security::test_input_sanitization("safe input").is_ok());
+        assert!(security::test_input_sanitization("normal task description").is_ok());
+    }
+
+    #[test]
+    fn test_output_format_injection() {
+        let harness = CliHarness::new();
+
+        // Test that format specifiers can't be abused
+        let malicious_formats = [
+            "human; rm -rf /",
+            "json && echo 'pwned'",
+            "yaml | cat /etc/passwd",
+        ];
+
+        for format in &malicious_formats {
+            harness
+                .execute(["--format", format, "config"])
+                .assert()
+                .failure();
+        }
+    }
+
+    #[test]
+    fn test_null_byte_injection() {
+        let harness = CliHarness::new();
+
+        // Test null byte injection attempts
+        let null_injected = "safe input\x00malicious".to_string();
+
+        let mut result = harness.execute(["episode", "create", &null_injected]);
+        match result.output() {
+            Ok(output) => {
+                // If command spawned, it should fail
+                assert!(
+                    output.status.code().is_some(),
+                    "Null byte injection should cause failure"
+                );
+            }
+            Err(_) => {
+                // Command failed to spawn due to null bytes - this is acceptable security behavior
+            }
+        }
+    }
+
+    #[test]
+    fn test_environment_variable_injection() {
+        let harness = CliHarness::new();
+
+        // Test environment variable injection attempts
+        let env_injected = "${HOME}/malicious";
+
+        harness
+            .execute(["episode", "create", env_injected])
+            .assert()
+            .failure();
+    }
+
+    #[test]
+    fn test_sql_injection_protection() {
+        let harness = CliHarness::new();
+
+        // Test SQL injection protection (parameterized queries)
+        let sql_injection_attempts = vec![
+            // Classic SQL injection
+            "task' OR '1'='1",
+            "task'; DROP TABLE episodes; --",
+            "task' UNION SELECT password FROM users--",
+            "task' AND 1=0 UNION SELECT username, password FROM admin--",
+            // Time-based injection attempts
+            "task' AND SLEEP(5)--",
+            "task' WAITFOR DELAY '0:0:5'--",
+            // Error-based injection
+            "task' AND 1=0 UNION SELECT 1,@@version--",
+            // Stacked queries
+            "task'; SELECT * FROM episodes; --",
+            "task'; DELETE FROM patterns WHERE 1=1; --",
+            // Comment evasion
+            "task'/*comment*/OR/*comment*/'1'='1",
+            "task'#comment\nOR '1'='1",
+        ];
+
+        for injection in sql_injection_attempts {
+            let mut result = harness.execute(["episode", "create", "--task", injection]);
+            let output = result.output().unwrap();
+
+            // Should fail safely without executing SQL
+            assert!(
+                output.status.code().is_some(),
+                "SQL injection '{}' caused crash",
+                injection
+            );
+
+            // Should not show signs of SQL execution (like actual data)
+            let _stdout = String::from_utf8_lossy(&output.stdout);
+            let _stderr = String::from_utf8_lossy(&output.stderr);
+
+            // The CLI echoes back the input, so it will contain SELECT/DROP if the input did.
+            // We rely on the fact that the process didn't crash and that proper parameterization is used internally.
+            // Naive string checks cause false positives.
+        }
+    }
+
+    #[test]
+    fn test_input_sanitization_comprehensive() {
+        let harness = CliHarness::new();
+
+        // Test comprehensive input sanitization with various attack vectors
+        let long_input = "x".repeat(10000);
+        let malicious_inputs = vec![
+            // Command injection
+            "; rm -rf /",
+            "&& echo 'hacked'",
+            "|| true",
+            "`whoami`",
+            "$(pwd)",
+            "${USER}",
+            "< /etc/passwd",
+            "> /dev/null",
+            "| cat",
+            "../etc/passwd",
+            "..\\windows\\system32",
+            // Template injection
+            "{{7*7}}",
+            "${config.__class__.__bases__[0].__subclasses__()}",
+            "{{config.__class__.__bases__[0].__subclasses__()}}",
+            // Path traversal
+            "../../../etc/passwd",
+            "..\\..\\..\\windows\\system32",
+            "/etc/passwd",
+            "C:\\windows\\system32",
+            "~/malicious",
+            "${HOME}/malicious",
+            "{{config.__class__.__bases__[0].__subclasses__()}}",
+            // Null byte injection
+            "task\x00malicious",
+            // Very long inputs (DoS attempt)
+            &long_input,
+        ];
+
+        for input in malicious_inputs {
+            // Test episode creation
+            let mut result = harness.execute(["episode", "create", "--task", input]);
+            let output = match result.output() {
+                Ok(output) => output,
+                Err(_) => {
+                    // Command failed to spawn (e.g., due to null bytes in arguments)
+                    // This is acceptable security behavior
+                    continue;
+                }
+            };
+
+            // Should fail safely, not execute dangerous operations
+            // The CLI may fail due to missing features, but shouldn't crash or execute commands
+            assert!(
+                output.status.code().is_some(),
+                "Input '{}' caused crash instead of safe failure",
+                input
+            );
+
+            // Should not contain signs of successful command execution
+            let _stdout = String::from_utf8_lossy(&output.stdout);
+            let _stderr = String::from_utf8_lossy(&output.stderr);
+        }
+    }
+
+    #[test]
+    fn test_output_format_injection_prevention() {
+        let harness = CliHarness::new();
+
+        // Test that output format specifiers can't be abused
+        let malicious_formats = vec![
+            "human; rm -rf /",
+            "json && echo 'hacked'",
+            "yaml | cat /etc/passwd",
+            "human' OR '1'='1",
+            "json<img src=x onerror=alert('xss')>",
+            "yaml${HOME}/malicious",
+        ];
+
+        for format in malicious_formats {
+            let mut result = harness.execute(["--format", format, "config"]);
+            let output = result.output().unwrap();
+
+            // Should fail safely
+            assert!(
+                output.status.code().is_some(),
+                "Format '{}' caused crash",
+                format
+            );
+
+            // Should not show signs of command execution
+            let stdout = String::from_utf8_lossy(&output.stdout);
+            let stderr = String::from_utf8_lossy(&output.stderr);
+
+            assert!(
+                !stdout.contains("hacked") && !stderr.contains("hacked"),
+                "Format '{}' may have led to command execution",
+                format
+            );
+        }
+    }
+}
